+++ frontend/src/page/body_fat.rs
@@ -29,7 +29,7 @@ pub fn init(
                 .keys()
                 .copied()
                 .collect::<Vec<NaiveDate>>(),
-            false,
+            common::DefaultInterval::_3M,
         ),
         dialog: Dialog::Hidden,
         loading: false,
@@ -418,7 +418,7 @@ pub fn update(
                             .keys()
                             .copied()
                             .collect::<Vec<NaiveDate>>(),
-                        false,
+                        common::DefaultInterval::_3M,
                     );
                 }
                 data::Event::BodyFatCreatedOk