+++ frontend/src/page/body_weight.rs
@@ -33,7 +33,7 @@ pub fn init(
                 .keys()
                 .copied()
                 .collect::<Vec<NaiveDate>>(),
-            false,
+            common::DefaultInterval::_3M,
         ),
         dialog: Dialog::Hidden,
         loading: false,
@@ -195,7 +195,7 @@ pub fn update(
                             .keys()
                             .copied()
                             .collect::<Vec<NaiveDate>>(),
-                        false,
+                        common::DefaultInterval::_3M,
                     );
                 }
                 data::Event::BodyWeightCreatedOk